Move archive utils to File Utils
