+++ components/apimgt/org.wso2.carbon.apimgt.rest.api.publisher/src/test/java/org/wso2/carbon/apimgt/rest/api/publisher/importexport/api/APIImportExportTestCase.java
@@ -21,25 +21,40 @@ package org.wso2.carbon.apimgt.rest.api.publisher.importexport.api;
 
 import org.apache.commons.collections.map.HashedMap;
 import org.apache.commons.io.IOUtils;
+import org.mockito.Mockito;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.testng.Assert;
 import org.testng.annotations.AfterClass;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
-import org.mockito.Mockito;
 import org.wso2.carbon.apimgt.core.api.APIPublisher;
 import org.wso2.carbon.apimgt.core.exception.APIManagementException;
-import org.wso2.carbon.apimgt.core.models.*;
+import org.wso2.carbon.apimgt.core.models.API;
+import org.wso2.carbon.apimgt.core.models.APIDetails;
+import org.wso2.carbon.apimgt.core.models.BusinessInformation;
+import org.wso2.carbon.apimgt.core.models.CorsConfiguration;
+import org.wso2.carbon.apimgt.core.models.DocumentContent;
+import org.wso2.carbon.apimgt.core.models.DocumentInfo;
+import org.wso2.carbon.apimgt.core.models.Endpoint;
 import org.wso2.carbon.apimgt.core.util.APIFileUtils;
 import org.wso2.carbon.apimgt.rest.api.publisher.dto.APIListDTO;
 import org.wso2.carbon.apimgt.rest.api.publisher.utils.ApiImportExportManager;
 import org.wso2.carbon.apimgt.rest.api.publisher.utils.FileBasedApiImportExportManager;
-import org.wso2.carbon.apimgt.rest.api.publisher.utils.ImportExportUtils;
 
-import java.io.*;
+import java.io.File;
+import java.io.FileInputStream;
+import java.io.IOException;
+import java.io.InputStream;
 import java.time.LocalDateTime;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
+import java.util.UUID;
 
 public class APIImportExportTestCase {
 
@@ -442,8 +457,8 @@ public class APIImportExportTestCase {
 
         // check if two APIs are written to the file system
         String unzipPath = importExportRootDirectory + File.separator + "unzipped-export-archive";
-        ImportExportUtils.extractArchive(exportedApiArchiveFilePath, unzipPath);
-        Assert.assertEquals(ImportExportUtils.getDirectoryList(unzipPath).size() == 2, true,
+        APIFileUtils.extractArchive(exportedApiArchiveFilePath, unzipPath);
+        Assert.assertEquals(APIFileUtils.getDirectoryList(unzipPath).size() == 2, true,
                 "Exported API count is not equal to 2");
 
         Mockito.when(apiPublisher.checkIfAPIExists(api2Id)).thenReturn(true);
